0906 상품페이지구현중

+++ mainProject/web/main/src/main/java/kr/co/test/vo/ProductVO.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class ProductVO {
-	private Long productId;
+	private int productId;
     private String productName;
-    private Long categoryId;
-    private Double price;
+    private int categoryId;
+    private int price;
     private Integer stockQuantity;
 }
